SWATCH, refs #1519: moved XmlGateKeeper from swatch::core to swatch::xml
